# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TrossApp API",
    "version": "1.0.0",
    "description": "\n        TrossApp Backend REST API - Skills-based work order management system.\n        \n        This documentation covers foundational authentication, user management, \n        and role-based access control endpoints.\n        \n        ## Authentication\n        \n        Most endpoints require a Bearer token in the Authorization header:\n        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n        \n        ## Dual Authentication Strategy\n        \n        - **Development Mode:** JWT-based authentication with test users\n        - **Production Mode:** Auth0 OAuth2/OIDC integration\n        \n        ## Response Format\n        \n        All successful responses follow this format:\n        ```json\n        {\n          \"success\": true,\n          \"data\": {...},\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n        \n        Error responses:\n        ```json\n        {\n          \"success\": false,\n          \"error\": \"Error message describing what went wrong\",\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n      ",
    "contact": {
      "name": "TrossApp Team",
      "email": "dev@trossapp.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
...
    },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/dev/token (dev) or Auth0 (prod)"
      }
    },
...
  },
...
  "paths": {
...
    "/api/dev/token": {
      "get": {
        "tags": [
          "Development"
        ],
        "summary": "Generate test token (any role)",
        "description": "Generate a JWT token for testing with any role.\n**DEVELOPMENT ONLY** - Always available regardless of AUTH_MODE.\nToken is valid for 24 hours.\n\nSupports all roles: admin, manager, dispatcher, technician, customer.\nDefaults to technician if no role specified (backward compatible).\n",
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "manager",
                "dispatcher",
                "technician",
                "customer"
              ]
            },
            "required": false,
            "description": "Role for the test token (defaults to technician)"
          }
        ],
        "responses": {
          "200": {
            "description": "Test token generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string",
                      "description": "JWT token for Authorization header"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "auth0_id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    },
                    "provider": {
                      "type": "string",
                      "example": "development"
                    },
                    "expires_in": {
                      "type": "integer",
                      "description": "Token validity in seconds"
                    },
                    "instructions": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid role specified"
          }
        }
      }
    },
...
  }
}